chore: auth/entities/model 하위에 있던 파일들 hooks, queries, stores로 분리
