+++ src/fsd_entities/auth/model/hooks/index.ts
@@ -1,0 +1,7 @@
+export * from './useLogin';
+export * from './useSubmitAcademicRecord';
+export * from './useResetPasswordForm';
+export * from './useFindEmailForm';
+export * from './useFindPasswordForm';
+export * from './useV2Form';
+export * from './useAcademicRecordForm';
