+++ src/fsd_entities/auth/model/queries/useSignUpForm.ts
@@ -4,7 +4,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
 import toast from 'react-hot-toast';
 
-import { signup } from '@/fsd_entities/auth/api/post';
+import { signup } from '../../api/post';
 import { parseErrorMessage } from '@/fsd_shared';
 
 const allowedKeys = [
